Fix reference import.
